PoC/requirements.txt
litellm>=1.40.0
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
uvicorn>=0.24.0
fastapi>=0.104.0
dnspython>=2.0.0
mcp>=0.1.0
prompt-toolkit>=3.0.0
filelock>=3.12.0
aiofiles>=23.0.0